+++ bluebird/api/resources/ic.py
@@ -7,7 +7,7 @@ from flask_restful import Resource, reqparse
 
 import bluebird.client as bb_client
 import bluebird.settings as settings
-from bluebird.api.resources.utils import wait_for_data
+from bluebird.api.resources.utils import check_ac_data
 
 PARSER = reqparse.RequestParser()
 PARSER.add_argument('filename', type=str, location='json', required=True)
@@ -48,13 +48,14 @@ class Ic(Resource):
 		start_paused = settings.SIM_MODE == 'agent'
 		err = bb_client.CLIENT_SIM.load_scenario(filename, speed=speed, start_paused=start_paused)
 
-		if not err:
-			resp = jsonify(f'Scenario {fn_base} loaded')
-			resp.status_code = 200
-		else:
+		if err:
 			resp = jsonify(f'Error: Could not load scenario {fn_base}. Error was: {err}')
 			resp.status_code = 500
 
-		wait_for_data()
+		err_resp = check_ac_data()
+		if err_resp:
+			return err_resp
 
+		resp = jsonify(f'Scenario {fn_base} loaded')
+		resp.status_code = 200
 		return resp