+++ bluebird/api/resources/loadlog.py
@@ -5,6 +5,8 @@ Provides logic for the Load Log API endpoint
 import logging
 import os
 
+import re
+import time
 import uuid
 from flask import jsonify
 from flask_restful import Resource, reqparse
@@ -12,8 +14,7 @@ from flask_restful import Resource, reqparse
 import bluebird.cache as bb_cache
 import bluebird.client as bb_client
 import bluebird.settings
-from bluebird.api.resources.utils import parse_lines, validate_scenario, wait_for_data, \
-	wait_for_pause
+from bluebird.api.resources.utils import check_ac_data, validate_scenario, wait_until_eq
 from bluebird.logging import store_local_scn
 from bluebird.utils.timeutils import timeit
 
@@ -23,6 +24,59 @@ PARSER = reqparse.RequestParser()
 PARSER.add_argument('filename', type=str, location='json', required=False)
 PARSER.add_argument('lines', type=str, location='json', required=False, action='append')
 PARSER.add_argument('time', type=int, location='json', required=True)
+
+
+def parse_lines(lines, target_time=0):
+	"""
+	Parses the content of an episode file
+	:param lines:
+	:param target_time:
+	:return: String for errors or dict with parsed data
+	"""
+
+	if not lines:
+		return 'No lines provided'
+
+	lines = list(lines)  # Take a copy of the input
+
+	# Get the seed from the first line
+	match = re.match(r'.*Episode started.*Seed is (\d+)', lines.pop(0))
+	if not match:
+		return 'Episode seed was not set'
+	scn_data = {'seed': int(match.group(1)), 'lines': []}
+
+	if not lines:
+		return 'No more lines after parsing seed'
+
+	while not 'Scenario file loaded' in lines[0]:
+		lines.pop(0)
+		if not lines:
+			return 'Couldn\'t find scenario content'
+
+	lines.pop(0)
+
+	while lines:
+		match = re.match(r'.*E.*>.*', lines[0])
+		if not match:
+			break
+		scn_data['lines'].append(lines.pop(0).split(' E ')[1])
+
+	if not lines:
+		return 'No more lines after parsing scenario content'
+
+	while lines:
+		line = lines.pop(0)
+		if 'Episode finished' in line:
+			return scn_data
+		match = re.match(r'.*C \[(\d+)\] (.*)$', line)
+		if match and not 'STEP' in line:
+			time_s = int(match.group(1))
+			if not target_time or time_s <= target_time:
+				cmd_time = time.strftime('%H:%M:%S', time.gmtime(time_s))
+				cmd_str = match.group(2)
+				scn_data['lines'].append(f'{cmd_time}> {cmd_str}')
+
+	return 'Could not find end of episode'
 
 
 class LoadLog(Resource):
@@ -50,8 +104,8 @@ class LoadLog(Resource):
 			return resp
 
 		target_time = parsed['time']
-		if target_time < 0:
-			resp = jsonify('Target time must be positive')
+		if target_time <= 0:
+			resp = jsonify('Target time must be greater than 0')
 			resp.status_code = 400
 			return resp
 
@@ -83,6 +137,15 @@ class LoadLog(Resource):
 			resp.status_code = 400
 			return resp
 
+		# Assert that the requested time is not past the end of the log
+		last_data = next(x for x in reversed(lines) if re.match(r'.*A \[(\d+)\] (.*)$', x))
+		last_time = int(re.search(r'\[(.*)]', last_data).group(1))
+
+		if target_time > last_time:
+			resp = jsonify(f'Error: Target time was greater than the latest time in the log')
+			resp.status_code = 400
+			return resp
+
 		err = validate_scenario(parsed_scn['lines'])
 
 		if err:
@@ -111,38 +174,46 @@ class LoadLog(Resource):
 			resp = jsonify(f'Error uploading scenario: {err}')
 			resp.status_code = 500
 
-		_LOGGER.debug('Starting the new scenario')
+		_LOGGER.info('Starting the new scenario')
 		err = bb_client.CLIENT_SIM.load_scenario(scn_name, start_paused=True)
 
 		if err:
-			resp = jsonify('Could not start scenario after upload')
+			resp = jsonify(f'Could not start scenario after upload: {err}')
 			resp.status_code = 500
 			return resp
 
 		_LOGGER.debug('Waiting for simulation to be paused')
-		wait_for_pause()
+		if not wait_until_eq(bb_cache.SIM_STATE.sim_state, 1):
+			resp = jsonify(f'Could not pause simulation after starting new scenario')
+			resp.status_code = 500
+			return resp
 
 		diff = target_time - bb_cache.SIM_STATE.sim_t
+		if diff:
+			# Naive approach - set DTMULT to target, then STEP once...
+			_LOGGER.debug(f'Time difference is {diff}. Stepping to {target_time}')
+			err = bb_client.CLIENT_SIM.send_stack_cmd(f'DTMULT {diff}')
 
-		# Naive approach - set DTMULT to target, then STEP once...
-		_LOGGER.debug(f'Time difference is {diff}. Stepping to {target_time}')
-		err = bb_client.CLIENT_SIM.send_stack_cmd(f'DTMULT {diff}')
+			if err:
+				resp = jsonify(f'Could not change speed: {err}')
+				resp.status_code = 500
+				return resp
 
-		if err:
-			resp = jsonify(f'Could not change speed: {err}')
-			resp.status_code = 500
-			return resp
+			_LOGGER.debug('Performing step')
+			err = bb_client.CLIENT_SIM.step()
 
-		_LOGGER.debug('Performing step')
-		err = bb_client.CLIENT_SIM.step()
-
-		if err:
-			resp = jsonify(f'Could not step simulations: {err}')
-			resp.status_code = 500
-			return resp
+			if err:
+				resp = jsonify(f'Could not step simulation: {err}')
+				resp.status_code = 500
+				return resp
+		else:
+			_LOGGER.debug(f"Simulation already at required time")
 
 		_LOGGER.debug('Waiting for AC_DATA to catch up')
-		wait_for_data()
+		err_resp = check_ac_data()
+		if err_resp:
+			return err_resp
+
 		bb_cache.AC_DATA.log()
 
 		# Reset DTMULT to the previous value