+++ tests/integration/test_episode_reload.py
@@ -3,13 +3,13 @@ Tests for the episode reloading
 """
 
 import requests
-import time
 from pyproj import Geod
 
 from tests.integration import API_URL_BASE
 
 _WGS84 = Geod(ellps='WGS84')
 _ONE_NM = 1852  # Meters
+_ONE_FT = 0.3048  # Meters
 
 
 def test_episode_reload_basic():
@@ -17,6 +17,8 @@ def test_episode_reload_basic():
 	Tests the basic functionality of the episode reloading
 	:return:
 	"""
+
+	test_acid = 'KL204'
 
 	resp = requests.post(f'{API_URL_BASE}/simmode', json={'mode': 'agent'})
 	assert resp.status_code == 200, 'Expected the mode to be set'
@@ -30,38 +32,44 @@ def test_episode_reload_basic():
 	resp = requests.post(f'{API_URL_BASE}/dtmult', json={'multiplier': 20})
 	assert resp.status_code == 200, 'Expected DTMULT to be set'
 
+	# Do an initial step
+	resp = requests.post(f'{API_URL_BASE}/step')
+	assert resp.status_code == 200, 'Expected the simulation was stepped'
+
+	# Change altitude and heading
+	resp = requests.post(f'{API_URL_BASE}/alt', json={'acid': test_acid, 'alt': 'FL100'})
+	assert resp.status_code == 200, 'Expected the aircraft altitude to be changed'
+
+	resp = requests.post(f'{API_URL_BASE}/hdg', json={'acid': test_acid, 'hdg': '180'})
+	assert resp.status_code == 200, 'Expected the aircraft heading to be changed'
+
 	for _ in range(3):
 		resp = requests.post(f'{API_URL_BASE}/step')
 		assert resp.status_code == 200, 'Expected the simulation was stepped'
 
-	test_acid = 'KL204'
-
+	# Get the position here
 	resp = requests.get(f'{API_URL_BASE}/pos?acid={test_acid}')
 	assert resp.status_code == 200, 'Expected to get the aircraft position'
 
 	initial_t = resp.json()['sim_t']
 	initial_pos = resp.json()[test_acid]
 
-	resp = requests.post(f'{API_URL_BASE}/alt', json={'acid': test_acid, 'alt': 'FL100'})
-	assert resp.status_code == 200, 'Expected to get the aircraft position'
-
-	for _ in range(5):
+	# Step some more
+	for _ in range(3):
 		resp = requests.post(f'{API_URL_BASE}/step')
 		assert resp.status_code == 200, 'Expected the simulation was stepped'
+
+	# Now get the episode log and reload to the middle of the log
 
 	resp = requests.get(f'{API_URL_BASE}/eplog')
 	assert resp.status_code == 200, 'Expected to receive the episode log'
 
 	episode_file = resp.json()['cur_ep_file']
-	target_t = 60
-
-	data = {'filename': episode_file, 'time': target_t}
-
-	start = time.time()
+	data = {'filename': episode_file, 'time': initial_t}
 	resp = requests.post(f'{API_URL_BASE}/loadlog', json=data)
-	print(f'\n! Time for reload: {time.time() - start}')
-
 	assert resp.status_code == 200, 'Expected the simulation was reloaded'
+
+	# Now get the reloaded position and compare with the initial
 
 	resp = requests.get(f'{API_URL_BASE}/pos?acid={test_acid}')
 	assert resp.status_code == 200, 'Expected to get the aircraft position'
@@ -69,15 +77,15 @@ def test_episode_reload_basic():
 	reloaded_t = resp.json()['sim_t']
 	reloaded_pos = resp.json()[test_acid]
 
-	assert reloaded_t == target_t, 'Expected the reloaded time to be at the target'
+	assert abs(reloaded_t - initial_t) <= 1, 'Expected the reloaded time to be at the target'
 
 	_, _, horizontal_sep_m = _WGS84.inv(initial_pos['lon'], initial_pos['lat'],
 	                                    reloaded_pos['lon'], reloaded_pos['lat'])
-	horizontal_sep_nm = round(horizontal_sep_m / _ONE_NM)
 
 	# TODO Check the deltas are reasonable...
-	assert horizontal_sep_nm <= 10, 'Expected positions to roughly match'
-	assert abs(reloaded_pos['alt'] - initial_pos['alt']) <= 100 / _ONE_NM, \
+	assert round(horizontal_sep_m / _ONE_NM) <= 5, \
+		'Expected positions to roughly match'
+	assert abs(reloaded_pos['alt'] - initial_pos['alt']) <= 100 / _ONE_FT, \
 		'Expected altitudes to roughly match'
 	assert abs(reloaded_pos['gs'] - initial_pos['gs']) <= 50, \
 		'Expected ground speeds to roughly match'