+++ tests/integration/test_metrics.py
@@ -19,6 +19,6 @@ def test_metrics_api_basic():
 	assert resp.status_code == 200, 'Expected the scenario to be loaded'
 
 	resp = requests.get(
-					f'{API_URL_BASE}/metric?name=vertical_separation&args=SCN1001,SCN1001')
+					f'{API_URL_BASE}/metric?name=aircraft_separation&args=SCN1001,SCN1001')
 	assert resp.status_code == 200, 'Expected valid args to return 200'
-	assert resp.json()['vertical_separation'] == -1, 'Expected -1'
+	assert resp.json()['aircraft_separation'] == -1, 'Expected -1'