+++ tests/unit/test_api_commands.py
@@ -256,3 +256,31 @@ def test_set_seed(client, patch_client_sim):
 	resp = client.post(API_PREFIX + '/seed', json=data)
 	assert resp.status == '200 OK'
 	assert bb.CLIENT_SIM.seed == 123, ''
+
+	data = {'value': 2 ** 32 - 1}
+	resp = client.post(API_PREFIX + '/seed', json=data)
+	assert resp.status == '200 OK'
+
+	data = {'value': 2 ** 32}
+	resp = client.post(API_PREFIX + '/seed', json=data)
+	assert resp.status == '400 BAD REQUEST'
+
+
+def test_alt_fl_parsing(client, patch_client_sim):
+	"""
+	Tests that we correctly parse the altitude argument
+	:return:
+	"""
+
+	acid = TEST_ACIDS[0]
+	data = {'acid': acid, 'alt': 'FL120'}
+
+	resp = client.post(API_PREFIX + '/alt', json=data)
+	assert resp.status_code == 200, 'Expected OK'
+	assert bb.CLIENT_SIM.last_stack_cmd == 'ALT TST1001 FL120', 'Expected ALT to match'
+
+	data['alt'] = '12345'
+
+	resp = client.post(API_PREFIX + '/alt', json=data)
+	assert resp.status_code == 200, 'Expected OK'
+	assert bb.CLIENT_SIM.last_stack_cmd == 'ALT TST1001 12345', 'Expected ALT to match'